jsongpt-client/jsongpt_client/models/ref_json_schema.py
from typing import Any, Dict, List, Type, TypeVar

import attr

T = TypeVar("T", bound="RefJsonSchema")


@attr.s(auto_attribs=True)
class RefJsonSchema:
    """
    Attributes:
        ref (str):
    """

    ref: str
    additional_properties: Dict[str, Any] = attr.ib(init=False, factory=dict)

    def to_dict(self) -> Dict[str, Any]:
        ref = self.ref

        field_dict: Dict[str, Any] = {}
        field_dict.update(self.additional_properties)
        field_dict.update(
            {
                "$ref": ref,
            }
        )

        return field_dict

    @classmethod
    def from_dict(cls: Type[T], src_dict: Dict[str, Any]) -> T:
        d = src_dict.copy()
        ref = d.pop("$ref")

        ref_json_schema = cls(
            ref=ref,
        )

        ref_json_schema.additional_properties = d
        return ref_json_schema

    @property
    def additional_keys(self) -> List[str]:
        return list(self.additional_properties.keys())

    def __getitem__(self, key: str) -> Any:
        return self.additional_properties[key]

    def __setitem__(self, key: str, value: Any) -> None:
        self.additional_properties[key] = value

    def __delitem__(self, key: str) -> None:
        del self.additional_properties[key]

    def __contains__(self, key: str) -> bool:
        return key in self.additional_properties
